🧪 Add quest success message test and update docs
