+++ frontend/e2e/backlog/docs-search.spec.ts
@@ -1,0 +1,6 @@
+import { test } from '@playwright/test';
+
+// TODO: Add docs search journey test.
+test('docs search flow placeholder', async () => {
+    // Placeholder test
+});
